+++ lib/layrz_models.freezed.dart
@@ -12,7 +12,7 @@ part of 'layrz_models.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
 
 MonitorRealWaypoint _$MonitorRealWaypointFromJson(Map<String, dynamic> json) {
   return _MonitorRealWaypoint.fromJson(json);
@@ -198,7 +198,7 @@ class _$MonitorRealWaypointImpl implements _MonitorRealWaypoint {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$MonitorRealWaypointImpl &&
@@ -543,7 +543,7 @@ class _$MonitorActiveCheckpointImpl implements _MonitorActiveCheckpoint {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$MonitorActiveCheckpointImpl &&
@@ -837,7 +837,7 @@ class _$LinkShortcutImpl implements _LinkShortcut {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LinkShortcutImpl &&
@@ -1150,7 +1150,7 @@ class _$AccessImpl implements _Access {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AccessImpl &&
@@ -1668,7 +1668,7 @@ class _$ActionImpl implements _Action {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ActionImpl &&
@@ -2038,7 +2038,7 @@ class _$ActionGeofenceSettingsImpl implements _ActionGeofenceSettings {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ActionGeofenceSettingsImpl &&
@@ -2786,7 +2786,7 @@ class _$AlgorithmImpl implements _Algorithm {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AlgorithmImpl &&
@@ -4357,7 +4357,7 @@ class _$AssetImpl implements _Asset {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AssetImpl &&
@@ -4904,7 +4904,7 @@ class _$ContactInfoImpl implements _ContactInfo {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ContactInfoImpl &&
@@ -5086,7 +5086,7 @@ class _$AssetLoginInfoImpl implements _AssetLoginInfo {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AssetLoginInfoImpl &&
@@ -5275,7 +5275,7 @@ class _$StaticPositionImpl implements _StaticPosition {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$StaticPositionImpl &&
@@ -5448,7 +5448,7 @@ class _$ZonePointImpl implements _ZonePoint {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ZonePointImpl &&
@@ -5673,7 +5673,7 @@ class _$AvatarImpl implements _Avatar {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AvatarImpl &&
@@ -6354,7 +6354,7 @@ class _$BillingPlanImpl implements _BillingPlan {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$BillingPlanImpl &&
@@ -6811,7 +6811,7 @@ class _$CareProtocolImpl implements _CareProtocol {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CareProtocolImpl &&
@@ -7115,7 +7115,7 @@ class _$CareTaskImpl implements _CareTask {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CareTaskImpl &&
@@ -7338,7 +7338,7 @@ class _$CommentOwnerImpl implements _CommentOwner {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CommentOwnerImpl &&
@@ -7558,7 +7558,7 @@ class _$CaseCommentImpl implements _CaseComment {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CaseCommentImpl &&
@@ -7973,7 +7973,7 @@ class _$CaseImpl implements _Case {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CaseImpl &&
@@ -8231,7 +8231,7 @@ class _$CategoryImpl implements _Category {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CategoryImpl &&
@@ -8559,7 +8559,7 @@ class _$LayrzChartImpl implements _LayrzChart {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LayrzChartImpl &&
@@ -8907,7 +8907,7 @@ class _$CheckpointImpl implements _Checkpoint {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CheckpointImpl &&
@@ -9193,7 +9193,7 @@ class _$WaypointImpl implements _Waypoint {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$WaypointImpl &&
@@ -9383,7 +9383,7 @@ class _$LinkedAssetToCheckpointIdImpl implements _LinkedAssetToCheckpointId {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LinkedAssetToCheckpointIdImpl &&
@@ -9983,7 +9983,7 @@ class _$CloudEntryImpl implements _CloudEntry {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CloudEntryImpl &&
@@ -10115,8 +10115,17 @@ mixin _$DeviceCommand {
   /// Is the [protocol], [protocolId] and the [model], [modelId] of the command.
   /// Only will comes when the command is a "template"
   String? get protocolId => throw _privateConstructorUsedError;
+
+  /// Is the [protocol], [protocolId] and the [model], [modelId] of the command.
+  /// Only will comes when the command is a "template"
   InboundProtocol? get protocol => throw _privateConstructorUsedError;
+
+  /// Is the [protocol], [protocolId] and the [model], [modelId] of the command.
+  /// Only will comes when the command is a "template"
   String? get modelId => throw _privateConstructorUsedError;
+
+  /// Is the [protocol], [protocolId] and the [model], [modelId] of the command.
+  /// Only will comes when the command is a "template"
   Model? get model => throw _privateConstructorUsedError;
 
   /// Is The command to execute.
@@ -10154,6 +10163,15 @@ mixin _$DeviceCommand {
   /// List of possible devices that can perform this command.
   List<Device>? get possibleDevices => throw _privateConstructorUsedError;
 
+  /// [modbusParameter] refers to the modbus parameter of the command.
+  /// This parameter contains the composition of the modbus command, like the controller address, function code,
+  /// register address, etc.
+  ModbusParameter? get modbusParameter => throw _privateConstructorUsedError;
+
+  /// [modbusPort] refers to the modbus port of the command.
+  /// This parameter contains the port number of the modbus command.
+  String? get modbusPort => throw _privateConstructorUsedError;
+
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
   $DeviceCommandCopyWith<DeviceCommand> get copyWith =>
@@ -10186,11 +10204,14 @@ abstract class $DeviceCommandCopyWith<$Res> {
       ExternalAccount? externalAccount,
       String? externalAccountId,
       List<Access>? access,
-      List<Device>? possibleDevices});
+      List<Device>? possibleDevices,
+      ModbusParameter? modbusParameter,
+      String? modbusPort});
 
   $InboundProtocolCopyWith<$Res>? get protocol;
   $ModelCopyWith<$Res>? get model;
   $ExternalAccountCopyWith<$Res>? get externalAccount;
+  $ModbusParameterCopyWith<$Res>? get modbusParameter;
 }
 
 /// @nodoc
@@ -10226,6 +10247,8 @@ class _$DeviceCommandCopyWithImpl<$Res, $Val extends DeviceCommand>
     Object? externalAccountId = freezed,
     Object? access = freezed,
     Object? possibleDevices = freezed,
+    Object? modbusParameter = freezed,
+    Object? modbusPort = freezed,
   }) {
     return _then(_value.copyWith(
       id: null == id
@@ -10308,6 +10331,14 @@ class _$DeviceCommandCopyWithImpl<$Res, $Val extends DeviceCommand>
           ? _value.possibleDevices
           : possibleDevices // ignore: cast_nullable_to_non_nullable
               as List<Device>?,
+      modbusParameter: freezed == modbusParameter
+          ? _value.modbusParameter
+          : modbusParameter // ignore: cast_nullable_to_non_nullable
+              as ModbusParameter?,
+      modbusPort: freezed == modbusPort
+          ? _value.modbusPort
+          : modbusPort // ignore: cast_nullable_to_non_nullable
+              as String?,
     ) as $Val);
   }
 
@@ -10346,6 +10377,18 @@ class _$DeviceCommandCopyWithImpl<$Res, $Val extends DeviceCommand>
       return _then(_value.copyWith(externalAccount: value) as $Val);
     });
   }
+
+  @override
+  @pragma('vm:prefer-inline')
+  $ModbusParameterCopyWith<$Res>? get modbusParameter {
+    if (_value.modbusParameter == null) {
+      return null;
+    }
+
+    return $ModbusParameterCopyWith<$Res>(_value.modbusParameter!, (value) {
+      return _then(_value.copyWith(modbusParameter: value) as $Val);
+    });
+  }
 }
 
 /// @nodoc
@@ -10376,7 +10419,9 @@ abstract class _$$DeviceCommandImplCopyWith<$Res>
       ExternalAccount? externalAccount,
       String? externalAccountId,
       List<Access>? access,
-      List<Device>? possibleDevices});
+      List<Device>? possibleDevices,
+      ModbusParameter? modbusParameter,
+      String? modbusPort});
 
   @override
   $InboundProtocolCopyWith<$Res>? get protocol;
@@ -10384,6 +10429,8 @@ abstract class _$$DeviceCommandImplCopyWith<$Res>
   $ModelCopyWith<$Res>? get model;
   @override
   $ExternalAccountCopyWith<$Res>? get externalAccount;
+  @override
+  $ModbusParameterCopyWith<$Res>? get modbusParameter;
 }
 
 /// @nodoc
@@ -10417,6 +10464,8 @@ class __$$DeviceCommandImplCopyWithImpl<$Res>
     Object? externalAccountId = freezed,
     Object? access = freezed,
     Object? possibleDevices = freezed,
+    Object? modbusParameter = freezed,
+    Object? modbusPort = freezed,
   }) {
     return _then(_$DeviceCommandImpl(
       id: null == id
@@ -10499,6 +10548,14 @@ class __$$DeviceCommandImplCopyWithImpl<$Res>
           ? _value._possibleDevices
           : possibleDevices // ignore: cast_nullable_to_non_nullable
               as List<Device>?,
+      modbusParameter: freezed == modbusParameter
+          ? _value.modbusParameter
+          : modbusParameter // ignore: cast_nullable_to_non_nullable
+              as ModbusParameter?,
+      modbusPort: freezed == modbusPort
+          ? _value.modbusPort
+          : modbusPort // ignore: cast_nullable_to_non_nullable
+              as String?,
     ));
   }
 }
@@ -10526,7 +10583,9 @@ class _$DeviceCommandImpl implements _DeviceCommand {
       this.externalAccount,
       this.externalAccountId,
       final List<Access>? access,
-      final List<Device>? possibleDevices})
+      final List<Device>? possibleDevices,
+      this.modbusParameter,
+      this.modbusPort})
       : _access = access,
         _possibleDevices = possibleDevices;
 
@@ -10558,10 +10617,19 @@ class _$DeviceCommandImpl implements _DeviceCommand {
   /// Only will comes when the command is a "template"
   @override
   final String? protocolId;
+
+  /// Is the [protocol], [protocolId] and the [model], [modelId] of the command.
+  /// Only will comes when the command is a "template"
   @override
   final InboundProtocol? protocol;
+
+  /// Is the [protocol], [protocolId] and the [model], [modelId] of the command.
+  /// Only will comes when the command is a "template"
   @override
   final String? modelId;
+
+  /// Is the [protocol], [protocolId] and the [model], [modelId] of the command.
+  /// Only will comes when the command is a "template"
   @override
   final Model? model;
 
@@ -10629,13 +10697,24 @@ class _$DeviceCommandImpl implements _DeviceCommand {
     return EqualUnmodifiableListView(value);
   }
 
+  /// [modbusParameter] refers to the modbus parameter of the command.
+  /// This parameter contains the composition of the modbus command, like the controller address, function code,
+  /// register address, etc.
+  @override
+  final ModbusParameter? modbusParameter;
+
+  /// [modbusPort] refers to the modbus port of the command.
+  /// This parameter contains the port number of the modbus command.
+  @override
+  final String? modbusPort;
+
   @override
   String toString() {
-    return 'DeviceCommand(id: $id, name: $name, source: $source, tagId: $tagId, deviceId: $deviceId, protocolId: $protocolId, protocol: $protocol, modelId: $modelId, model: $model, command: $command, devicePassword: $devicePassword, payload: $payload, commandId: $commandId, isHexCoded: $isHexCoded, username: $username, scriptName: $scriptName, externalAccount: $externalAccount, externalAccountId: $externalAccountId, access: $access, possibleDevices: $possibleDevices)';
+    return 'DeviceCommand(id: $id, name: $name, source: $source, tagId: $tagId, deviceId: $deviceId, protocolId: $protocolId, protocol: $protocol, modelId: $modelId, model: $model, command: $command, devicePassword: $devicePassword, payload: $payload, commandId: $commandId, isHexCoded: $isHexCoded, username: $username, scriptName: $scriptName, externalAccount: $externalAccount, externalAccountId: $externalAccountId, access: $access, possibleDevices: $possibleDevices, modbusParameter: $modbusParameter, modbusPort: $modbusPort)';
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$DeviceCommandImpl &&
@@ -10669,7 +10748,11 @@ class _$DeviceCommandImpl implements _DeviceCommand {
                 other.externalAccountId == externalAccountId) &&
             const DeepCollectionEquality().equals(other._access, _access) &&
             const DeepCollectionEquality()
-                .equals(other._possibleDevices, _possibleDevices));
+                .equals(other._possibleDevices, _possibleDevices) &&
+            (identical(other.modbusParameter, modbusParameter) ||
+                other.modbusParameter == modbusParameter) &&
+            (identical(other.modbusPort, modbusPort) ||
+                other.modbusPort == modbusPort));
   }
 
   @JsonKey(ignore: true)
@@ -10695,7 +10778,9 @@ class _$DeviceCommandImpl implements _DeviceCommand {
         externalAccount,
         externalAccountId,
         const DeepCollectionEquality().hash(_access),
-        const DeepCollectionEquality().hash(_possibleDevices)
+        const DeepCollectionEquality().hash(_possibleDevices),
+        modbusParameter,
+        modbusPort
       ]);
 
   @JsonKey(ignore: true)
@@ -10733,7 +10818,9 @@ abstract class _DeviceCommand implements DeviceCommand {
       final ExternalAccount? externalAccount,
       final String? externalAccountId,
       final List<Access>? access,
-      final List<Device>? possibleDevices}) = _$DeviceCommandImpl;
+      final List<Device>? possibleDevices,
+      final ModbusParameter? modbusParameter,
+      final String? modbusPort}) = _$DeviceCommandImpl;
 
   factory _DeviceCommand.fromJson(Map<String, dynamic> json) =
       _$DeviceCommandImpl.fromJson;
@@ -10764,10 +10851,19 @@ abstract class _DeviceCommand implements DeviceCommand {
   /// Only will comes when the command is a "template"
   String? get protocolId;
   @override
+
+  /// Is the [protocol], [protocolId] and the [model], [modelId] of the command.
+  /// Only will comes when the command is a "template"
   InboundProtocol? get protocol;
   @override
+
+  /// Is the [protocol], [protocolId] and the [model], [modelId] of the command.
+  /// Only will comes when the command is a "template"
   String? get modelId;
   @override
+
+  /// Is the [protocol], [protocolId] and the [model], [modelId] of the command.
+  /// Only will comes when the command is a "template"
   Model? get model;
   @override
 
@@ -10815,6 +10911,17 @@ abstract class _DeviceCommand implements DeviceCommand {
 
   /// List of possible devices that can perform this command.
   List<Device>? get possibleDevices;
+  @override
+
+  /// [modbusParameter] refers to the modbus parameter of the command.
+  /// This parameter contains the composition of the modbus command, like the controller address, function code,
+  /// register address, etc.
+  ModbusParameter? get modbusParameter;
+  @override
+
+  /// [modbusPort] refers to the modbus port of the command.
+  /// This parameter contains the port number of the modbus command.
+  String? get modbusPort;
   @override
   @JsonKey(ignore: true)
   _$$DeviceCommandImplCopyWith<_$DeviceCommandImpl> get copyWith =>
@@ -10947,7 +11054,7 @@ class _$ConnectionImpl implements _Connection {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ConnectionImpl &&
@@ -11160,7 +11267,7 @@ class _$CountryImpl implements _Country {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CountryImpl &&
@@ -11337,7 +11444,7 @@ class _$PhoneNumberImpl extends _PhoneNumber {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$PhoneNumberImpl &&
@@ -11505,7 +11612,7 @@ class _$NullablePhoneNumberImpl extends _NullablePhoneNumber {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$NullablePhoneNumberImpl &&
@@ -11717,7 +11824,7 @@ class _$CityImpl implements _City {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CityImpl &&
@@ -11916,7 +12023,7 @@ class _$CountryStateImpl implements _CountryState {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CountryStateImpl &&
@@ -12320,7 +12427,7 @@ class _$CredentialFieldImpl implements _CredentialField {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CredentialFieldImpl &&
@@ -12615,7 +12722,7 @@ class _$CustomFieldImpl implements _CustomField {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CustomFieldImpl &&
@@ -12800,7 +12907,7 @@ class _$CustomFieldInputImpl implements _CustomFieldInput {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CustomFieldInputImpl &&
@@ -13099,7 +13206,7 @@ class _$DbPartitionImpl implements _DbPartition {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$DbPartitionImpl &&
@@ -13336,7 +13443,7 @@ class _$DepartmentImpl implements _Department {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$DepartmentImpl &&
@@ -13945,7 +14052,7 @@ class _$DeviceImpl implements _Device {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$DeviceImpl &&
@@ -14304,7 +14411,7 @@ class _$EmailTemplateImpl implements _EmailTemplate {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$EmailTemplateImpl &&
@@ -14784,7 +14891,7 @@ class _$EmployeeImpl implements _Employee {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$EmployeeImpl &&
@@ -15064,7 +15171,7 @@ class _$FtpAccountImpl implements _FtpAccount {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$FtpAccountImpl &&
@@ -15615,7 +15722,7 @@ class _$LayrzFunctionImpl implements _LayrzFunction {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LayrzFunctionImpl &&
@@ -15974,7 +16081,7 @@ class _$GenericPermissionItemImpl implements _GenericPermissionItem {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GenericPermissionItemImpl &&
@@ -16498,7 +16605,7 @@ class _$GenericPermissionImpl implements _GenericPermission {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GenericPermissionImpl &&
@@ -16713,7 +16820,7 @@ class _$GeofencePointImpl implements _GeofencePoint {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GeofencePointImpl &&
@@ -17249,7 +17356,7 @@ class _$GeofenceImpl implements _Geofence {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GeofenceImpl &&
@@ -17852,7 +17959,7 @@ class _$GeoPointImpl implements _GeoPoint {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GeoPointImpl &&
@@ -18060,7 +18167,7 @@ class _$GuideImpl implements _Guide {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$GuideImpl &&
@@ -18244,7 +18351,7 @@ class _$LanguageImpl implements _Language {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LanguageImpl &&
@@ -18474,7 +18581,7 @@ class _$LintErrorImpl implements _LintError {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LintErrorImpl &&
@@ -18764,7 +18871,7 @@ class _$ModelImpl implements _Model {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ModelImpl &&
@@ -19443,7 +19550,7 @@ class _$OperationImpl implements _Operation {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$OperationImpl &&
@@ -19766,7 +19873,7 @@ class _$HttpHeaderImpl implements _HttpHeader {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$HttpHeaderImpl &&
@@ -20704,7 +20811,7 @@ class _$PresetImpl implements _Preset {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$PresetImpl &&
@@ -20976,7 +21083,7 @@ class _$ProfileImpl implements _Profile {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ProfileImpl &&
@@ -21260,7 +21367,7 @@ class _$ReferenceImpl implements _Reference {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ReferenceImpl &&
@@ -21637,7 +21744,7 @@ class _$ReportTemplateImpl implements _ReportTemplate {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ReportTemplateImpl &&
@@ -21929,7 +22036,7 @@ class _$ReportTemplatePageImpl implements _ReportTemplatePage {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ReportTemplatePageImpl &&
@@ -22156,7 +22263,7 @@ class _$ReportTemplateColImpl implements _ReportTemplateCol {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ReportTemplateColImpl &&
@@ -22364,7 +22471,7 @@ class _$CustomReportImpl implements _CustomReport {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$CustomReportImpl &&
@@ -23291,7 +23398,7 @@ class _$SensorImpl implements _Sensor {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$SensorImpl &&
@@ -23761,7 +23868,7 @@ class _$MaskPointImpl implements _MaskPoint {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$MaskPointImpl &&
@@ -23940,7 +24047,7 @@ class _$SensorPairImpl implements _SensorPair {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$SensorPairImpl &&
@@ -24193,7 +24300,7 @@ class _$SuspendedServiceImpl implements _SuspendedService {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$SuspendedServiceImpl &&
@@ -25821,7 +25928,7 @@ class _$TagImpl implements _Tag {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TagImpl &&
@@ -26460,7 +26567,7 @@ class _$DeviceTelemetryImpl extends _DeviceTelemetry {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$DeviceTelemetryImpl &&
@@ -26805,7 +26912,7 @@ class _$AssetTelemetryImpl extends _AssetTelemetry {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$AssetTelemetryImpl &&
@@ -27142,7 +27249,7 @@ class _$TelemetryPositionImpl extends _TelemetryPosition {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TelemetryPositionImpl &&
@@ -27354,7 +27461,7 @@ class _$TelemetrySensorImpl extends _TelemetrySensor {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TelemetrySensorImpl &&
@@ -27580,7 +27687,7 @@ class _$TimezoneImpl implements _Timezone {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TimezoneImpl &&
@@ -27752,7 +27859,7 @@ class _$TokenImpl implements _Token {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TokenImpl &&
@@ -28645,7 +28752,7 @@ class _$TriggerImpl implements _Trigger {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TriggerImpl &&
@@ -29223,7 +29330,7 @@ class _$TriggerActivationImpl implements _TriggerActivation {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TriggerActivationImpl &&
@@ -30353,7 +30460,7 @@ class _$TripImpl extends _Trip {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$TripImpl &&
@@ -30611,7 +30718,7 @@ class _$PointImpl extends _Point {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$PointImpl &&
@@ -30801,7 +30908,7 @@ class _$ConciergeFormImpl implements _ConciergeForm {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$ConciergeFormImpl &&
@@ -31820,7 +31927,7 @@ class _$UserImpl implements _User {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$UserImpl &&
@@ -32314,7 +32421,7 @@ class _$InviteLinkImpl extends _InviteLink {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$InviteLinkImpl &&